updated CI

+++ .ci/install.sh
@@ -10,10 +10,12 @@ sudo apt-mark hold msodbcsql17 mssql-tools # microsoft wants to manually accept 
 sudo apt-get -y install grub-efi
 sudo update-grub
 
+if [ "$distro" = "20.04" ]; then
 # the "gce-compute-image-packages" package often freezes the installation at some point
 # the installation freezes when it tries to manage some systemd services
 # this attempts to install the package and stop the problematic service during the process
 ((sleep 90 && (sudo systemctl stop google-instance-setup.service && echo "gce service stoped" || echo "gce service not stoped")) & (sudo timeout 120s apt-get -y install gce-compute-image-packages)) || echo "\e[1;31mInstallation of gce-compute-image-packages failed\e[0m"
+fi
 
 sudo apt-get -y upgrade --fix-missing
